DOCS: rename frontend files with kebab case instead of snake case
