+++ frontend/src/features/settings-finder/api/settings-finder.js
@@ -1,6 +1,6 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 
-import { uploadImage } from "@/services/api_settings";
+import { uploadImage } from "@/services/api-image-settings";
 
 export const useSettingsMutation = ({ setError }) => {
   const queryClient = useQueryClient();
