+++ frontend/src/services/api-ai-model.js
