+++ frontend/src/services/api-csam-image.js
