+++ frontend/src/services/api-image-settings.js
